# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BattingSummary": {
        "properties": {
          "average": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average"
          },
          "balls": {
            "title": "Balls",
            "type": "integer"
          },
          "boundary_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary Pct"
          },
          "dot_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dot Pct"
          },
          "ducks": {
            "title": "Ducks",
            "type": "integer"
          },
          "fifties": {
            "title": "Fifties",
            "type": "integer"
          },
          "fours": {
            "title": "Fours",
            "type": "integer"
          },
          "highest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HighScore"
              },
              {
                "type": "null"
              }
            ]
          },
          "hundreds": {
            "title": "Hundreds",
            "type": "integer"
          },
          "innings": {
            "title": "Innings",
            "type": "integer"
          },
          "matches": {
            "title": "Matches",
            "type": "integer"
          },
          "not_outs": {
            "title": "Not Outs",
            "type": "integer"
          },
          "outs": {
            "title": "Outs",
            "type": "integer"
          },
          "par_average": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Par Average"
          },
          "par_boundary_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Par Boundary Pct"
          },
          "par_dot_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Par Dot Pct"
          },
          "par_strike_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Par Strike Rate"
          },
          "runs": {
            "title": "Runs",
            "type": "integer"
          },
          "runs_above_par": {
            "description": "Runs scored minus par runs for the same balls.",
            "title": "Runs Above Par",
            "type": "number"
          },
          "sixes": {
            "title": "Sixes",
            "type": "integer"
          },
          "strike_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike Rate"
          },
          "wpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Win probability added, in probability units.",
            "title": "Wpa"
          },
          "wpa_innings": {
            "description": "Innings with win probabilities behind ``wpa``.",
            "title": "Wpa Innings",
            "type": "integer"
          }
        },
        "required": [
          "matches",
          "innings",
          "not_outs",
          "runs",
          "balls",
          "outs",
          "average",
          "strike_rate",
          "highest",
          "fifties",
          "hundreds",
          "ducks",
          "fours",
          "sixes",
          "dot_pct",
          "boundary_pct",
          "par_strike_rate",
          "par_average",
          "par_dot_pct",
          "par_boundary_pct",
          "runs_above_par",
          "wpa",
          "wpa_innings"
        ],
        "title": "BattingSummary",
        "type": "object"
      },
...
      "BowlingSummary": {
        "properties": {
          "average": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average"
          },
          "balls": {
            "title": "Balls",
            "type": "integer"
          },
          "best": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BestFigures"
              },
              {
                "type": "null"
              }
            ]
          },
          "boundary_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary Pct"
          },
          "dot_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dot Pct"
          },
          "economy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Economy"
          },
          "five_wickets": {
            "description": "Innings with five or more wickets.",
            "title": "Five Wickets",
            "type": "integer"
          },
          "four_wickets": {
            "description": "Innings with exactly four wickets.",
            "title": "Four Wickets",
            "type": "integer"
          },
          "innings": {
            "title": "Innings",
            "type": "integer"
          },
          "maidens": {
            "title": "Maidens",
            "type": "integer"
          },
          "matches": {
            "title": "Matches",
            "type": "integer"
          },
          "noballs": {
            "title": "Noballs",
            "type": "integer"
          },
          "overs": {
            "title": "Overs",
            "type": "string"
          },
          "par_dot_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Par Dot Pct"
          },
          "par_economy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Par Economy"
          },
          "par_strike_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Par Strike Rate"
          },
          "runs": {
            "title": "Runs",
            "type": "integer"
          },
          "runs_saved": {
            "description": "Par runs minus runs conceded for the same balls.",
            "title": "Runs Saved",
            "type": "number"
          },
          "strike_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike Rate"
          },
          "wickets": {
            "title": "Wickets",
            "type": "integer"
          },
          "wides": {
            "title": "Wides",
            "type": "integer"
          },
          "wpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wpa"
          },
          "wpa_innings": {
            "title": "Wpa Innings",
            "type": "integer"
          }
        },
        "required": [
          "matches",
          "innings",
          "balls",
          "overs",
          "runs",
          "wickets",
          "average",
          "economy",
          "strike_rate",
          "best",
          "four_wickets",
          "five_wickets",
          "maidens",
          "dot_pct",
          "boundary_pct",
          "wides",
          "noballs",
          "par_economy",
          "par_strike_rate",
          "par_dot_pct",
          "runs_saved",
          "wpa",
          "wpa_innings"
        ],
        "title": "BowlingSummary",
        "type": "object"
      },
...
      "Dismissals": {
        "properties": {
          "batting": {
            "description": "How the player got out.",
            "items": {
              "$ref": "#/components/schemas/DismissalCount"
            },
            "title": "Batting",
            "type": "array"
          },
          "bowling": {
            "description": "How the player took wickets.",
            "items": {
              "$ref": "#/components/schemas/DismissalCount"
            },
            "title": "Bowling",
            "type": "array"
          }
        },
        "required": [
          "batting",
          "bowling"
        ],
        "title": "Dismissals",
        "type": "object"
      },
...
      "FieldingSummary": {
        "properties": {
          "catches": {
            "title": "Catches",
            "type": "integer"
          },
          "run_outs": {
            "title": "Run Outs",
            "type": "integer"
          },
          "stumpings": {
            "title": "Stumpings",
            "type": "integer"
          }
        },
        "required": [
          "catches",
          "stumpings",
          "run_outs"
        ],
        "title": "FieldingSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Percentiles": {
        "properties": {
          "batting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PercentileGroup"
              },
              {
                "type": "null"
              }
            ]
          },
          "bowling": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PercentileGroup"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "batting",
          "bowling"
        ],
        "title": "Percentiles",
        "type": "object"
      },
...
      "PhaseSplits": {
        "properties": {
          "batting": {
            "items": {
              "$ref": "#/components/schemas/PhaseBatting"
            },
            "title": "Batting",
            "type": "array"
          },
          "bowling": {
            "items": {
              "$ref": "#/components/schemas/PhaseBowling"
            },
            "title": "Bowling",
            "type": "array"
          }
        },
        "required": [
          "batting",
          "bowling"
        ],
        "title": "PhaseSplits",
        "type": "object"
      },
      "PlayerBio": {
        "properties": {
          "batting_hand": {
            "anyOf": [
              {
                "enum": [
                  "right",
                  "left"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batting Hand"
          },
          "bowling_arm": {
            "anyOf": [
              {
                "enum": [
                  "right",
                  "left"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bowling Arm"
          },
          "bowling_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bowling Style"
          },
          "bowling_type": {
            "anyOf": [
              {
                "enum": [
                  "pace",
                  "spin"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bowling Type"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "first_season": {
            "title": "First Season",
            "type": "integer"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "is_keeper": {
            "title": "Is Keeper",
            "type": "boolean"
          },
          "last_season": {
            "title": "Last Season",
            "type": "integer"
          },
          "matches": {
            "title": "Matches",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "player_id": {
            "title": "Player Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "batter",
              "bowler",
              "all_rounder"
            ],
            "title": "Role",
            "type": "string"
          },
          "teams": {
            "description": "Franchises played for, most recent first.",
            "items": {
              "$ref": "#/components/schemas/TeamStint"
            },
            "title": "Teams",
            "type": "array"
          }
        },
        "required": [
          "player_id",
          "name",
          "full_name",
          "country",
          "date_of_birth",
          "batting_hand",
          "bowling_arm",
          "bowling_type",
          "bowling_style",
          "role",
          "is_keeper",
          "first_season",
          "last_season",
          "matches",
          "teams"
        ],
        "title": "PlayerBio",
        "type": "object"
      },
      "PlayerListItem": {
        "properties": {
          "batting_average": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batting Average"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "economy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Economy"
          },
          "first_season": {
            "description": "Career debut season.",
            "title": "First Season",
            "type": "integer"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "is_keeper": {
            "title": "Is Keeper",
            "type": "boolean"
          },
          "last_season": {
            "description": "Most recent season played.",
            "title": "Last Season",
            "type": "integer"
          },
          "matches": {
            "description": "Matches within the filters.",
            "title": "Matches",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "player_id": {
            "title": "Player Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "batter",
              "bowler",
              "all_rounder"
            ],
            "title": "Role",
            "type": "string"
          },
          "runs": {
            "title": "Runs",
            "type": "integer"
          },
          "strike_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike Rate"
          },
          "team": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeamTag"
              },
              {
                "type": "null"
              }
            ],
            "description": "Most recent team within the filters."
          },
          "wickets": {
            "title": "Wickets",
            "type": "integer"
          }
        },
        "required": [
          "player_id",
          "name",
          "full_name",
          "country",
          "role",
          "is_keeper",
          "team",
          "first_season",
          "last_season",
          "matches",
          "runs",
          "batting_average",
          "strike_rate",
          "wickets",
          "economy"
        ],
        "title": "PlayerListItem",
        "type": "object"
      },
      "PlayerPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlayerListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PlayerPage",
        "type": "object"
      },
      "PlayerProfile": {
        "properties": {
          "batting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BattingSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "bowling": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BowlingSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "dismissals": {
            "$ref": "#/components/schemas/Dismissals"
          },
          "fielding": {
            "$ref": "#/components/schemas/FieldingSummary"
          },
          "percentiles": {
            "$ref": "#/components/schemas/Percentiles"
          },
          "phases": {
            "$ref": "#/components/schemas/PhaseSplits"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerBio"
          },
          "recent": {
            "$ref": "#/components/schemas/RecentInnings"
          },
          "seasons": {
            "items": {
              "$ref": "#/components/schemas/SeasonLine"
            },
            "title": "Seasons",
            "type": "array"
          },
          "window": {
            "$ref": "#/components/schemas/SeasonWindow"
          }
        },
        "required": [
          "player",
          "window",
          "batting",
          "bowling",
          "fielding",
          "seasons",
          "phases",
          "percentiles",
          "recent",
          "dismissals"
        ],
        "title": "PlayerProfile",
        "type": "object"
      },
...
      "PlayerSplits": {
        "properties": {
          "batting": {
            "items": {
              "$ref": "#/components/schemas/BattingSplitGroup"
            },
            "title": "Batting",
            "type": "array"
          },
          "bowling": {
            "items": {
              "$ref": "#/components/schemas/BowlingSplitGroup"
            },
            "title": "Bowling",
            "type": "array"
          },
          "window": {
            "$ref": "#/components/schemas/SeasonWindow"
          }
        },
        "required": [
          "window",
          "batting",
          "bowling"
        ],
        "title": "PlayerSplits",
        "type": "object"
      },
      "RecentInnings": {
        "properties": {
          "batting": {
            "items": {
              "$ref": "#/components/schemas/BattingInnings"
            },
            "title": "Batting",
            "type": "array"
          },
          "bowling": {
            "items": {
              "$ref": "#/components/schemas/BowlingInnings"
            },
            "title": "Bowling",
            "type": "array"
          }
        },
        "required": [
          "batting",
          "bowling"
        ],
        "title": "RecentInnings",
        "type": "object"
      },
...
      "SeasonLine": {
        "properties": {
          "batting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SeasonBatting"
              },
              {
                "type": "null"
              }
            ]
          },
          "bowling": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SeasonBowling"
              },
              {
                "type": "null"
              }
            ]
          },
          "matches": {
            "title": "Matches",
            "type": "integer"
          },
          "season": {
            "title": "Season",
            "type": "integer"
          },
          "teams": {
            "items": {
              "type": "string"
            },
            "title": "Teams",
            "type": "array"
          }
        },
        "required": [
          "season",
          "teams",
          "matches",
          "batting",
          "bowling"
        ],
        "title": "SeasonLine",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Cricket intelligence, ball by ball. Model outputs are estimates, not guarantees of real outcomes.",
    "title": "CricIQ API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/players": {
      "get": {
        "description": "Player directory with headline numbers; season and team filters scope the numbers.",
        "operationId": "list_players_api_v1_players_get",
        "parameters": [
          {
            "description": "Name search",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Name search",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "batter",
                    "bowler",
                    "all_rounder",
                    "keeper"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2100,
                  "minimum": 2008,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "description": "Franchise id, e.g. MI",
            "in": "query",
            "name": "team",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 8,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Franchise id, e.g. MI",
              "title": "Team"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "matches",
              "enum": [
                "matches",
                "runs",
                "wickets",
                "recent",
                "name"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Players",
        "tags": [
          "players"
        ]
      }
    },
    "/api/v1/players/{player_id}": {
      "get": {
        "description": "Profile: batting and bowling against par, seasons, phases, percentiles and recent form.",
        "operationId": "read_player_api_v1_players__player_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "string"
            }
          },
          {
            "description": "First season (inclusive).",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2100,
                  "minimum": 2008,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First season (inclusive).",
              "title": "From"
            }
          },
          {
            "description": "Last season (inclusive).",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2100,
                  "minimum": 2008,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last season (inclusive).",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Player",
        "tags": [
          "players"
        ]
      }
    },
    "/api/v1/players/{player_id}/splits": {
      "get": {
        "description": "Batting and bowling split by phase, opponent type, position, venue, opposition and more.",
        "operationId": "read_splits_api_v1_players__player_id__splits_get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "string"
            }
          },
          {
            "description": "First season (inclusive).",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2100,
                  "minimum": 2008,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First season (inclusive).",
              "title": "From"
            }
          },
          {
            "description": "Last season (inclusive).",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2100,
                  "minimum": 2008,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last season (inclusive).",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSplits"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Splits",
        "tags": [
          "players"
        ]
      }
    },
...
  }
}